# packages/api-client/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/admin/audit": {
      "get": {
        "operationId": "AdminOrgController_audit_v1",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "integer",
              "minimum": 1,
              "maximum": 10000
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "default": 25,
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "action",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "maxLength": 64
            }
          },
          {
            "name": "entityType",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "maxLength": 64
            }
          },
          {
            "name": "entityId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "maxLength": 64
            }
          },
          {
            "name": "actor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date",
              "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))$"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date",
              "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditList"
                }
              }
            }
          }
        },
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/settings": {
      "get": {
        "operationId": "AdminOrgController_settings_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminOrganization"
                }
              }
            }
          }
        },
        "tags": [
          "admin"
        ]
      },
      "put": {
        "operationId": "AdminOrgController_updateSettings_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrganizationSettingsInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminOrganization"
                }
              }
            }
          }
        },
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/branches": {
      "get": {
        "operationId": "AdminBranchesController_list_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminBranchList"
                }
              }
            }
          }
        },
        "tags": [
          "admin: branches"
        ]
      },
      "post": {
        "operationId": "AdminBranchesController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BranchInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminBranchDetail"
                }
              }
            }
          }
        },
        "tags": [
          "admin: branches"
        ]
      }
    },
    "/api/v1/admin/branches/{slug}": {
      "get": {
        "operationId": "AdminBranchesController_get_v1",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 80,
              "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminBranchDetail"
                }
              }
            }
          }
        },
        "tags": [
          "admin: branches"
        ]
      },
      "put": {
        "operationId": "AdminBranchesController_update_v1",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 80,
              "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BranchInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminBranchDetail"
                }
              }
            }
          }
        },
        "tags": [
          "admin: branches"
        ]
      }
    },
    "/api/v1/admin/branches/{slug}/archive": {
      "post": {
        "operationId": "AdminBranchesController_archive_v1",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 80,
              "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminBranchDetail"
                }
              }
            }
          }
        },
        "summary": "Hide a branch from the site (its content stays in the database).",
        "tags": [
          "admin: branches"
        ]
      }
    },
    "/api/v1/admin/branches/{slug}/restore": {
      "post": {
        "operationId": "AdminBranchesController_restore_v1",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 80,
              "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminBranchDetail"
                }
              }
            }
          }
        },
        "tags": [
          "admin: branches"
        ]
      }
    },
    "/api/v1/admin/branches/{slug}/schedules": {
      "post": {
        "operationId": "AdminBranchesController_createSchedule_v1",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 80,
              "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminBranchDetail"
                }
              }
            }
          }
        },
        "tags": [
          "admin: branches"
        ]
      }
    },
    "/api/v1/admin/branches/{slug}/schedules/{id}": {
      "put": {
        "operationId": "AdminBranchesController_updateSchedule_v1",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 80,
              "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminBranchDetail"
                }
              }
            }
          }
        },
        "tags": [
          "admin: branches"
        ]
      },
      "delete": {
        "operationId": "AdminBranchesController_deleteSchedule_v1",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 80,
              "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminBranchDetail"
                }
              }
            }
          }
        },
        "tags": [
          "admin: branches"
        ]
      }
    },
    "/api/v1/admin/branches/{slug}/leaders": {
      "post": {
        "operationId": "AdminBranchesController_createLeader_v1",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 80,
              "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LeaderInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminBranchDetail"
                }
              }
            }
          }
        },
        "tags": [
          "admin: branches"
        ]
      }
    },
    "/api/v1/admin/branches/{slug}/leaders/{id}": {
      "put": {
        "operationId": "AdminBranchesController_updateLeader_v1",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 80,
              "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LeaderInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminBranchDetail"
                }
              }
            }
          }
        },
        "tags": [
          "admin: branches"
        ]
      },
      "delete": {
        "operationId": "AdminBranchesController_deleteLeader_v1",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 80,
              "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminBranchDetail"
                }
              }
            }
          }
        },
        "tags": [
          "admin: branches"
        ]
      }
    }
  },
  "info": {
    "title": "Church Platform API",
    "description": "REST API of the Church Platform. Authentication uses an HttpOnly session cookie; unsafe methods also require the CSRF token cookie echoed in the X-CSRF-Token header.",
    "version": "1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "cp_session"
      }
    },
    "schemas": {
...
      "AuditList": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuditEntry"
            }
          },
          "page": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "pageSize": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "total": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          }
        },
        "required": [
          "items",
          "page",
          "pageSize",
          "total"
        ],
        "additionalProperties": false
      },
...
      "AdminOrganization": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "shortName": {
            "nullable": true,
            "type": "string"
          },
          "tagline": {
            "nullable": true,
            "type": "string"
          },
          "description": {
            "nullable": true,
            "type": "string"
          },
          "email": {
            "nullable": true,
            "type": "string"
          },
          "phone": {
            "nullable": true,
            "type": "string"
          },
          "websiteUrl": {
            "nullable": true,
            "type": "string"
          },
          "timezone": {
            "type": "string"
          },
          "registrationOpen": {
            "type": "boolean"
          },
          "baptismRequestsEnabled": {
            "type": "boolean"
          },
          "defaultInAppCategories": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "ANNOUNCEMENTS",
                "UPDATES",
                "EVENTS",
                "NEWS",
                "SERMONS",
                "BAPTISM",
                "MEMBERSHIP",
                "ACCOUNT"
              ]
            }
          },
          "defaultEmailCategories": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "ANNOUNCEMENTS",
                "UPDATES",
                "EVENTS",
                "NEWS",
                "SERMONS",
                "BAPTISM",
                "MEMBERSHIP",
                "ACCOUNT"
              ]
            }
          },
          "socialLinks": {
            "type": "object",
            "properties": {
              "facebook": {
                "nullable": true,
                "type": "string"
              },
              "youtube": {
                "nullable": true,
                "type": "string"
              },
              "instagram": {
                "nullable": true,
                "type": "string"
              },
              "x": {
                "nullable": true,
                "type": "string"
              }
            },
            "required": [
              "facebook",
              "youtube",
              "instagram",
              "x"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "name",
          "shortName",
          "tagline",
          "description",
          "email",
          "phone",
          "websiteUrl",
          "timezone",
          "registrationOpen",
          "baptismRequestsEnabled",
          "defaultInAppCategories",
          "defaultEmailCategories",
          "socialLinks"
        ],
        "additionalProperties": false
      },
      "OrganizationSettingsInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 160
          },
          "shortName": {
            "nullable": true,
            "type": "string",
            "maxLength": 60
          },
          "tagline": {
            "nullable": true,
            "type": "string",
            "maxLength": 200
          },
          "description": {
            "nullable": true,
            "type": "string",
            "maxLength": 5000
          },
          "email": {
            "nullable": true,
            "type": "string",
            "maxLength": 254,
            "format": "email",
            "pattern": "^(?:[A-Za-z0-9_'+\\-]+\\.)*[A-Za-z0-9_'+\\-]*[A-Za-z0-9_+-]@(?:[A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
          },
          "phone": {
            "nullable": true,
            "type": "string",
            "maxLength": 40
          },
          "websiteUrl": {
            "nullable": true,
            "type": "string",
            "maxLength": 2048,
            "format": "uri"
          },
          "timezone": {
            "type": "string",
            "minLength": 1,
            "maxLength": 64
          },
          "registrationOpen": {
            "type": "boolean"
          },
          "baptismRequestsEnabled": {
            "type": "boolean"
          },
          "defaultInAppCategories": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "ANNOUNCEMENTS",
                "UPDATES",
                "EVENTS",
                "NEWS",
                "SERMONS",
                "BAPTISM",
                "MEMBERSHIP",
                "ACCOUNT"
              ]
            }
          },
          "defaultEmailCategories": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "ANNOUNCEMENTS",
                "UPDATES",
                "EVENTS",
                "NEWS",
                "SERMONS",
                "BAPTISM",
                "MEMBERSHIP",
                "ACCOUNT"
              ]
            }
          },
          "socialLinks": {
            "type": "object",
            "properties": {
              "facebook": {
                "nullable": true,
                "type": "string",
                "maxLength": 2048,
                "format": "uri"
              },
              "youtube": {
                "nullable": true,
                "type": "string",
                "maxLength": 2048,
                "format": "uri"
              },
              "instagram": {
                "nullable": true,
                "type": "string",
                "maxLength": 2048,
                "format": "uri"
              },
              "x": {
                "nullable": true,
                "type": "string",
                "maxLength": 2048,
                "format": "uri"
              }
            }
          }
        },
        "required": [
          "name",
          "timezone",
          "registrationOpen",
          "baptismRequestsEnabled",
          "defaultInAppCategories",
          "defaultEmailCategories",
          "socialLinks"
        ]
      },
      "AdminBranchList": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminBranchRow"
            }
          }
        },
        "required": [
          "items"
        ],
        "additionalProperties": false
      },
...
      "BranchInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "slug": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80,
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
          },
          "type": {
            "default": "MAIN",
            "type": "string",
            "enum": [
              "MAIN",
              "SUB",
              "SATELLITE",
              "ONLINE"
            ]
          },
          "parentBranch": {
            "nullable": true,
            "type": "string",
            "minLength": 1,
            "maxLength": 80,
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
          },
          "description": {
            "nullable": true,
            "type": "string",
            "maxLength": 5000
          },
          "addressLine1": {
            "nullable": true,
            "type": "string",
            "maxLength": 200
          },
          "addressLine2": {
            "nullable": true,
            "type": "string",
            "maxLength": 200
          },
          "city": {
            "nullable": true,
            "type": "string",
            "maxLength": 120
          },
          "province": {
            "nullable": true,
            "type": "string",
            "maxLength": 120
          },
          "postalCode": {
            "nullable": true,
            "type": "string",
            "maxLength": 20
          },
          "countryCode": {
            "default": "ZA",
            "type": "string",
            "pattern": "^[A-Z]{2}$"
          },
          "latitude": {
            "nullable": true,
            "type": "number",
            "minimum": -90,
            "maximum": 90
          },
          "longitude": {
            "nullable": true,
            "type": "number",
            "minimum": -180,
            "maximum": 180
          },
          "mapsUrl": {
            "nullable": true,
            "type": "string",
            "maxLength": 2048,
            "format": "uri"
          },
          "phone": {
            "nullable": true,
            "type": "string",
            "maxLength": 40
          },
          "email": {
            "nullable": true,
            "type": "string",
            "maxLength": 254,
            "format": "email",
            "pattern": "^(?:[A-Za-z0-9_'+\\-]+\\.)*[A-Za-z0-9_'+\\-]*[A-Za-z0-9_+-]@(?:[A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
          },
          "sortOrder": {
            "default": 0,
            "type": "integer",
            "minimum": 0,
            "maximum": 10000
          }
        },
        "required": [
          "name"
        ]
      },
...
      "LeaderInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "bio": {
            "nullable": true,
            "type": "string",
            "maxLength": 1000
          },
          "sortOrder": {
            "default": 0,
            "type": "integer",
            "minimum": 0,
            "maximum": 10000
          },
          "isActive": {
            "default": true,
            "type": "boolean"
          }
        },
        "required": [
          "name",
          "title"
        ]
      }
    }
  }
}